SERVICES-865 small refactoring
