+++ src/graphql/entities/account.detailed/account.detailed.module.ts
@@ -4,10 +4,9 @@ import { AccountDetailedResolver } from "src/graphql/entities/account.detailed/a
 import { AccountModule } from "src/endpoints/accounts/account.module";
 import { CollectionModule } from "src/endpoints/collections/collection.module";
 import { NftModule } from "src/endpoints/nfts/nft.module";
-import { TokenModule } from "src/endpoints/tokens/token.module";
 
 @Module({
-  imports: [AccountModule, CollectionModule, NftModule, TokenModule],
+  imports: [AccountModule, CollectionModule, NftModule],
   providers: [AccountDetailedResolver],
 })
 export class AccountDetailedModule { }
