+++ src/graphql/entities/tokens/tokens.input.ts
@@ -2,6 +2,7 @@ import { Field, Float, ID, InputType } from "@nestjs/graphql";
 import { SortOrder } from "src/common/entities/sort.order";
 import { TokenFilter } from "src/endpoints/tokens/entities/token.filter";
 import { TokenSort } from "src/endpoints/tokens/entities/token.sort";
+import { GetAccountDetailedInput } from "../account.detailed/account.detailed.input";
 
 @InputType({ description: "Input to retreive the given tokens count for." })
 export class GetTokensCountInput {
@@ -88,4 +89,33 @@ export class GetTokenAccountsInput extends GetTokenInput {
 
   @Field(() => Float, { name: "size", description: "Number of tokens to retrieve for the given result set.", nullable: true, defaultValue: 25 })
   size: number = 25;
+}
+
+@InputType({ description: "Input to retrieve the given account details with roles for." })
+export class GetAccountTokenRolesInput extends GetAccountDetailedInput {
+  constructor(partial?: Partial<GetAccountTokenRolesInput>) {
+    super();
+    Object.assign(this, partial);
+  }
+
+  @Field(() => Float, { name: "from", description: "Number of tokens to skip for the given result set.", nullable: true, defaultValue: 0 })
+  from: number = 0;
+
+  @Field(() => Float, { name: "size", description: "Number of tokens to retrieve for the given result set.", nullable: true, defaultValue: 25 })
+  size: number = 25;
+
+  @Field(() => ID, { name: "identifier", description: "Tokens identifier to retrieve for the given result set.", nullable: true })
+  identifier?: string;
+
+  @Field(() => String, { name: "search", description: "Tokens search to retrieve for the given result set.", nullable: true })
+  search?: string;
+
+  @Field(() => String, { name: "creator", description: "Owner to retrieve for the given result set.", nullable: true })
+  owner?: string;
+
+  @Field(() => Boolean, { name: "canMint", description: "Filter by property canMint to retrieve the given result set.", nullable: true })
+  canMint?: boolean;
+
+  @Field(() => Boolean, { name: "canBurn", description: "Filter by property canBurn to retrieve the given result set.", nullable: true })
+  canBurn?: boolean;
 }
